services/fastapi/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pillow==10.4.0
numpy==1.26.4
mlflow==2.16.0
python-multipart==0.0.12
tensorflow==2.15.0
prometheus-fastapi-instrumentator==6.1.0
